Phase 3: E2E validation, dashboard, benchmark expansion, onboarding, report
End-to-end validation:
- test_e2e_pipeline.py: 16 tests with real tiny model forward pass
  through extract → probes → SAE → attention → geometry
- scripts/dry_run.py: full pipeline on 5 items in ~5s on CPU

Benchmark expansion:
- 330 items (up from 284), 165 matched pairs across 7 categories
- All underrepresented categories now ≥20 pairs
- 0 CRITICAL audit issues

Interactive dashboard:
- 6-tab Gradio app (behavioral, probes, SAE, attention, geometry, hypotheses)
- --synthetic mode for demo without real results

Results aggregator:
- Unified report from per-workstream JSONs with H1-H6 hypothesis evaluation
- Markdown + JSON output, matches pre-registration thresholds exactly

GPU runbooks:
- First-day-on-GPU guide for FASRC/Kempner (6 phases) and RunPod
- deploy/preflight_check.py: automated environment verification

Team onboarding:
- 4-file onboarding suite: guide, background reading, 5 coding exercises, 20 FAQs
- All exercises work on synthetic data (no GPU needed)

Paper expansion:
- methods.tex: +170 lines (statistical methodology, confound controls)
- appendix.tex: +313 lines (benchmark details, probe specs, SAE protocol, compute budget)
- benchmark.tex: +60 lines (example items, scoring rubric)

Code quality:
- Lint fixes across 48 files (-261 lines of dead imports/code)
- F722 added to ruff ignores for jaxtyping multi-dim shapes

427/427 tests passing. Smoke test green.

+++ src/s1s2/sae/loaders.py
@@ -241,7 +241,7 @@ class MockSAE:
     ) -> Float[Tensor, "batch n_features"]:
         z = x.float() @ self.W_enc + self.b_enc
         if self.sparsity < 1.0:
-            k = max(1, int(round(self.sparsity * self.n_features)))
+            k = max(1, round(self.sparsity * self.n_features))
             # Keep top-k absolute magnitudes, zero the rest.
             topk = torch.topk(z.abs(), k=k, dim=-1).indices
             mask = torch.zeros_like(z)
